+++ docs/changelog.md
@@ -1,11 +1,15 @@
 # Changelog
 
+## 2024-10-17
+
+### Added
+
+- Support for fail-timeout, backends-path, proxy-cache-valid configuration options from integration with Content Cache Backends Config charm.
+
 ## 2024-10-07
 
 ### Added
 
 - Changelog for tracking user-relevant changes.
-
 - Basic content cache functionality with nginx.
-
-- Support for location, backends, protocol configuration options from integration with Content Cache Backends Config charm.
+- Support for hostname, path, backends, protocol configuration options from integration with Content Cache Backends Config charm.